requirements.txt
boto3==1.4.1
botocore==1.4.60
certbot==0.8.1
cffi==1.7.0
ConfigArgParse==0.10.0
configobj==5.0.6
cryptography==1.4
docutils==0.12
enum34==1.1.6
funcsigs==1.0.2
futures==3.0.5
idna==2.1
ipaddress==1.0.16
jmespath==0.9.0
lxml==3.6.4
mock==2.0.0
ndg-httpsclient==0.4.2
parsedatetime==2.1
pbr==1.10.0
psutil==4.3.0
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyOpenSSL==16.0.0
pyRFC3339==1.0
python-dateutil==2.5.3
python2-pythondialog==3.4.0
pytz==2016.6.1
requests==2.10.0
requests-file==1.4
route53==1.0.1
s3transfer==0.1.7
six==1.10.0
tldextract==2.0.1
zope.component==4.2.2
zope.event==4.2.0
zope.interface==4.2.0
